refactor: 사이드 바 공통 컴포넌트로 분리 #79
